Allow for async callback in rpc response handling
Currently, when a response is being handled, the callback must
be called before the before response function returns. This is
because the response returns immidiately, and doesn't wait for
the callback to be called.

To allow for async response handling, this simply waits for the
callback to be called before returning.

NW